fix(common): Handle downloads from sources that don’t provide a "Content-Length" header
In some cases, sources do not return the Content-Length header. This can happen due to a proxy or simply because the source does not include it by design.

Currently, you-get either:
- Attempts to download the next range with an invalid Range header (out of bounds), resulting in an HTTP error, or
- Gets stuck in an infinite loop of attempts.

This PR fixes the issue by adjusting the exit condition:
If the expected file size is infinite (i.e., undetermined) and the buffer has been fully read, the loop exits successfully.

Additionally, this PR includes other fixes to ensure proper functionality:
- Correctly processes URLs when provided as a list with a single element.
- Prevents deleting the temporary file in this case, as it is actually the final result, not a temporary file.

A test has been added to reproduce and verify the issue.

+++ src/you_get/common.py
@@ -672,6 +672,7 @@ def url_save(
     headers=None, timeout=None, **kwargs
 ):
     tmp_headers = headers.copy() if headers is not None else {}
+    url = url[0] if type(url) is list and len(url) == 1 else url
     # When a referer specified with param refer,
     # the key must be 'Referer' for the hack here
     if refer is not None:
@@ -807,9 +808,9 @@ def url_save(
                     except socket.timeout:
                         pass
                     if not buffer:
-                        if is_chunked and received_chunk == range_length:
+                        if is_chunked and (received_chunk == range_length or range_length == float('inf')):
                             break
-                        elif not is_chunked and received == file_size:  # Download finished
+                        elif not is_chunked and (received == file_size or range_length == float('inf')):  # Download finished
                             break
                         # Unexpected termination. Retry request
                         tmp_headers['Range'] = 'bytes=' + str(received - chunk_start) + '-'
@@ -827,10 +828,11 @@ def url_save(
         received, os.path.getsize(temp_filepath), temp_filepath
     )
 
-    if os.access(filepath, os.W_OK):
-        # on Windows rename could fail if destination filepath exists
-        os.remove(filepath)
-    os.rename(temp_filepath, filepath)
+    if temp_filepath != filepath:
+        if os.access(filepath, os.W_OK):
+            # on Windows rename could fail if destination filepath exists
+            os.remove(filepath)
+        os.rename(temp_filepath, filepath)
 
 
 class SimpleProgressBar:
